remove generic analyis type and move types to common dir
